Fixed bug in rendering of mixer (parameters were not being sampled correctly).

+++ ColaLib/ColaLib/CCOL/CCOLComponentMixer.cpp
@@ -37,7 +37,7 @@ void CCOLComponentMixer::renderOutputs(unsigned int numframes) {
     SignalType *outputBuffer = output->prepareBufferOfSize(numframes);
     
     for (int i = 0; i < numframes; i++) {
-        float delta = i / numframes;
+        float delta = i / (float)numframes;
         float level1Output = level1->getOutputAtDelta(delta);
         float level2Output = level2->getOutputAtDelta(delta);
